requirements.txt
pandas
numpy
ta
oandapyV20
python-dotenv
